fix(libc): warning

+++ src/libc/stdio/vsprintf.c
@@ -870,15 +870,15 @@ int _vsnprintf(char *buffer, const size_t maxlen, const char *format,
                 const bool is_ll = sizeof(uintptr_t) == sizeof(long long);
                 if (is_ll) {
                     idx = _ntoa_long_long(out, buffer, idx, maxlen,
-                                          (uintptr_t)va_arg(va, void *), false,
-                                          16U, precision, width, flags);
+                                          (unsigned long)va_arg(va, void *),
+                                          false, 16U, precision, width, flags);
                 }
                 else {
 #endif
                     idx = _ntoa_long(
                         out, buffer, idx, maxlen,
-                        (unsigned long)((uintptr_t)va_arg(va, void *)), false,
-                        16U, precision, width, flags);
+                        (unsigned long)((unsigned long)va_arg(va, void *)),
+                        false, 16U, precision, width, flags);
 #if defined(PRINTF_SUPPORT_LONG_LONG)
                 }
 #endif
